cmapplot global variables

Expose a list of predefined variables for use by the cmapplot package and its users. List includes commonly used colors, font and font size specifications, and a list of constants which aid in drawing cmap-themed plots.

cmapplot_globals

Format

An object of class list of length 8.

Plot Constants

cmapplot_globals$consts contains a list of default constants that set certain plot aesthetics. Units of all plot constants are "bigpts": 1/72 of an inch. Most plot constants are invoked (and can be overridden) in finalize_plot: these are marked F. Some are used/can be overridden in theme_cmap: these are marked with T.

lwd_strongline: This stronger-width line is drawn vertically or horizontally with the hline, vline args of theme_cmap(). (T)

lwd_gridline: This thinner-width line is drawn vertically or horizontally with the gridlines, axislines args of theme_cmap(). (T)

lwd_plotline: The width of any lines drawn by geoms in the plot (e.g. geom_line) but not explicitly sized by the geom's aesthetic. Implemented by finalize_plot or by apply_cmap_default_aes but not overridable in either context. (Modify by setting the size explicitly in the geom, but see gg_lwd_convert first.)

lwd_topline: The width of the line above the plot and title. (F)

margin_topline_t: The margin between the top edge of the image and the top line. (F)

length_ticks: The length of the axis ticks (if shown).(T)

margin_title_t: The margin between the top line and the title. (F)

margin_title_b: The margin between the title and the caption. (F)

margin_caption_b: The margin between the bottom of the (F) caption and the bottom edge of the image.

margin_legend_t: The margin between the top line and the plot box (i.e., the top of the legend). (F)

margin_legend_i: The margin between legends (this only applies in plots with two or more legends and does not affect legend spacing on plots with single legends that have multiple rows). (T, F)

margin_legend_b: The margin between the bottom of the legend and the rest of the plot. (T, F)

margin_plot_b: The margin between the bottom of the plot and the bottom edge of the image. (F)

margin_title_l: The margin between the left edge of the image and the title. This also applies to the caption. Deducted from title_width. (F)

margin_title_r: The margin between the right edge of the image and the title. This also applies to the caption. Deducted from title_width. (F)

margin_plot_r: The margin between the right edge of the plot and the edge of the image. (F)

margin_panel_r: Padding between the plot and its right-hand drawing extent. Override this based on space needed for x axis labels. (T)

leading_title: Text leading for Title text. (F)

leading_caption: Text leading for Caption text. (F)